requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.1.0=pypi_0
addict=2.4.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=3.6.1=pypi_0
appdirs=1.4.4=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pypi_0
attrdict=2.0.1=pypi_0
attrs=21.4.0=pypi_0
babel=2.10.3=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
bleach=5.0.1=pypi_0
blis=0.7.8=pypi_0
boto3=1.24.21=pypi_0
botocore=1.27.21=pypi_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.2.0=pypi_0
catalogue=1.0.0=pypi_0
certifi=2022.6.15=pypi_0
cffi=1.15.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
cryptography=37.0.1=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
cymem=2.0.6=pypi_0
debugpy=1.6.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
deprecation=2.1.0=pypi_0
docker-pycreds=0.4.0=pypi_0
easydict=1.9=pypi_0
entrypoints=0.4=pypi_0
executing=0.8.3=pypi_0
fastjsonschema=2.15.3=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.7.1=pypi_0
flit-core=3.8.0=py38h06a4308_0
fonttools=4.33.3=pypi_0
freetype=2.12.1=h4a9f257_0
fvcore=0.1.5.post20220512=pypi_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.9.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.47.0=pypi_0
h5py=3.7.0=pypi_0
huggingface-hub=0.8.1=pypi_0
hydra-core=1.2.0=pypi_0
idna=3.3=pypi_0
imageio=2.19.3=pypi_0
importlib-metadata=4.12.0=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.9=pypi_0
ipykernel=6.15.0=pypi_0
ipython=8.4.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.1=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.2=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.1.0=pypi_0
jpeg=9e=h5eee18b_1
json5=0.9.8=pypi_0
jsonlines=3.1.0=pypi_0
jsonschema=4.6.1=pypi_0
jupyter-client=7.3.4=pypi_0
jupyter-core=4.10.0=pypi_0
jupyter-packaging=0.12.2=pypi_0
jupyter-server=1.18.1=pypi_0
jupyterlab=3.4.3=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.15.0=pypi_0
jupyterlab-widgets=1.1.1=pypi_0
kiwisolver=1.4.3=pypi_0
lame=3.100=h7b6447c_0
langcodes=3.3.0=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lmdb=1.3.0=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.3.7=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.2=pypi_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mmcv=1.5.3=pypi_0
murmurhash=1.0.7=pypi_0
nbclassic=0.4.2=pypi_0
nbclient=0.6.6=pypi_0
nbconvert=6.5.0=pypi_0
nbformat=5.4.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.5=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=pypi_0
nltk=3.7=pypi_0
notebook=6.4.12=pypi_0
notebook-shim=0.1.0=pypi_0
numpy=1.22.3=pypi_0
oauthlib=3.2.0=pypi_0
omegaconf=2.2.2=pypi_0
open3d=0.15.2=pypi_0
opencv-python=4.6.0.66=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.4.3=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathtools=0.1.2=pypi_0
pathy=0.6.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=pypi_0
pip=21.2.4=pypi_0
plac=1.1.3=pypi_0
plyfile=0.7.4=pypi_0
pointnet2=3.0.0=pypi_0
portalocker=2.4.0=pypi_0
preshed=3.0.6=pypi_0
prometheus-client=0.14.1=pypi_0
prompt-toolkit=3.0.30=pypi_0
protobuf=3.19.4=pypi_0
psutil=5.9.1=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocoevalcap=1.2=pypi_0
pycocotools=2.0.4=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.8.2=pypi_0
pygments=2.12.0=pypi_0
pyopenssl=22.0.0=pypi_0
pyparsing=3.0.9=pypi_0
pyquaternion=0.9.9=pypi_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.16=h7a1cb2a_3
python-dateutil=2.8.2=pypi_0
pytorch=1.12.0=py3.8_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytorch-transformers=1.2.0=pypi_0
pytz=2022.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=23.2.0=pypi_0
readline=8.2=h5eee18b_0
regex=2022.6.2=pypi_0
requests=2.28.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
s3transfer=0.6.0=pypi_0
sacremoses=0.0.53=pypi_0
scikit-learn=1.1.1=pypi_0
scipy=1.8.1=pypi_0
send2trash=1.8.0=pypi_0
sentencepiece=0.1.96=pypi_0
sentry-sdk=1.21.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=61.2.0=pypi_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
smart-open=5.2.1=pypi_0
smmap=5.0.0=pypi_0
sniffio=1.2.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
spacy=2.3.7=pypi_0
spacy-legacy=3.0.9=pypi_0
spacy-loggers=1.0.2=pypi_0
sqlite=3.41.2=h5eee18b_0
srsly=1.0.5=pypi_0
stack-data=0.3.0=pypi_0
tabulate=0.8.10=pypi_0
tensorboard=2.9.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.5.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.15.0=pypi_0
thinc=7.4.5=pypi_0
threadpoolctl=3.1.0=pypi_0
timm=0.6.5=pypi_0
tinycss2=1.1.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.12.1=pypi_0
tomlkit=0.11.0=pypi_0
torchaudio=0.12.0=py38_cu113
torchvision=0.13.0=py38_cu113
tornado=6.2=pypi_0
tqdm=4.64.0=pypi_0
traitlets=5.3.0=pypi_0
transformers=4.20.1=pypi_0
trimesh=2.35.39=pypi_0
typer=0.4.2=pypi_0
typing-extensions=4.1.1=pypi_0
urllib3=1.26.15=py38h06a4308_0
wandb=0.15.0=pypi_0
warmup-scheduler=0.3=pypi_0
wasabi=0.9.1=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.3.3=pypi_0
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pypi_0
widgetsnbextension=3.6.1=pypi_0
xz=5.2.10=h5eee18b_1
yacs=0.1.8=pypi_0
yapf=0.32.0=pypi_0
zipp=3.8.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
